data_analysis/requirements.txt
pandas
streamlit
matplotlib
faker
mysql-connector-python
Pillow
passlib==1.7.4
python-jose
geopandas
